# Replace mutex in config service with atomic config entry / semaphore