--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,7 +7,7 @@ repository = "https://github.com/configcat/rust-sdk"
 documentation = "https://configcat.com/docs/sdk-reference/rust"
 keywords = ["configcat", "feature-flag", "feature-toggle"]
 license = "MIT"
-version = "0.1.2"
+version = "0.1.3"
 edition = "2021"
 
 [dependencies]
@@ -24,6 +24,7 @@ sha1 = "0.10"
 sha2 = "0.10"
 base16ct = { version = "0.2", features = ["alloc"] }
 semver = "1.0"
+arc-swap = "1.7.1"
 
 [dev-dependencies]
 mockito = "1.2.0"