--- a/benches/client_benchmark.rs
+++ b/benches/client_benchmark.rs
@@ -42,7 +42,7 @@ fn get_value_bench(c: &mut Criterion) {
             for _ in 0..200 {
                 let cl = client.clone();
                 handles.push(tokio::spawn(async move {
-                    cl.get_value("isAwesomeFeatureEnabled", false, None).await
+                    cl.get_value("testKey", false, None).await
                 }));
             }
             for handle in handles {