### Describe the purpose of your pull request
This PR refines the synchronization mechanism in `ConfigService`. Rather than a `Mutex`, we now use atomic operations on the in-memory config data and `Semaphore` to ensure only one HTTP call is active at a time.

### Related issues (only if applicable)
- #8 

### Requirement checklist (only if applicable)
- [x] I have covered the applied changes with automated tests.
- [x] I have executed the full automated test set against my changes.
- [x] I have validated my changes against all supported platform versions.
- [x] I have read and accepted the [contribution agreement](https://github.com/configcat/legal/blob/main/contribution-agreement.md).