@chamons 2025-06-16
I do wonder, if you increase the number of concurrent readers does the time go up exponentially or faster than linearly?