@z4kn4fein 2025-06-16
@chamons I've added a steady increase in the reader count to the benchmark, these are the results:

Old:
```
100    time:   [145.67 µs 150.86 µs 155.88 µs]
1000   time:   [1.2553 ms 1.2826 ms 1.3099 ms]
10000  time:   [13.828 ms 14.247 ms 14.705 ms]
100000 time:   [218.45 ms 230.06 ms 241.82 ms]
```

New:
```
100    time:   [43.322 µs 44.457 µs 45.863 µs]
1000   time:   [445.83 µs 455.46 µs 466.94 µs]
10000  time:   [5.8118 ms 5.9303 ms 6.0545 ms]
100000 time:   [89.342 ms 90.955 ms 92.680 ms]
```

The execution times still increase faster than linearly but with a lower amount.